+++ MOSES_FC/dma.h
@@ -52,7 +52,7 @@ void sort(roeimage_t*);
 void unsort(roeimage_t * image);
 int dmaClearBlock();
 void dmaClose();
-int rotr(short);
+short rotr(short);
 
 #endif	/* DMA_H */
